Removing old articles
remove_old_articles function takes a database file and a site ID as arguments and removes the oldest articles from the articles table for the given site until there are only 15 articles left. It does this by first querying the database to count the number of articles for the given site, then deleting the appropriate number of rows from the table using an SQL DELETE statement.